move testCatalog to the 'common' folder
